chore(test): move `TestEnv` and tests to `tests/`
